django_project/requirements.txt
django >= 5.2.1
django-filter >= 2.5.1
djangorestframework >= 3.16.0
Markdown >= 3.8